error message work, help icon, add button styles/animations, some state cleanup (wip)
